requirements.txt
tensorflow-datasets==3.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.8.0
Pillow